docs(changeset): split package into vanilla/react/formatter entry points
