+++ tests/atom-logger.test.ts
